move privateRecordsChoice helper from Increase to All Claims
