+++ src/applications/disability-benefits/526EZ/constants.js
@@ -1,4 +1,3 @@
-export const VA_FORM4142_URL = 'https://www.vba.va.gov/pubs/forms/VBA-21-4142-ARE.pdf';
 export const E_BENEFITS_URL = 'https://www.ebenefits.va.gov/ebenefits/homepage';
 
 export const itfStatuses = {
